+++ lib/legion/extensions/llm/ledger/transport/transport.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/transport'
-require 'legion/llm/transport/exchanges/metering'
-require 'legion/llm/transport/exchanges/audit'
+require_relative 'exchanges/metering'
+require_relative 'exchanges/audit'
 
 module Legion
   module Extensions
@@ -14,17 +14,17 @@ module Legion
           def self.additional_e_to_q
             [
               {
-                from:        Legion::LLM::Transport::Exchanges::Metering,
+                from:        Legion::Extensions::Llm::Ledger::Transport::Exchanges::Metering,
                 to:          Legion::Extensions::Llm::Ledger::Transport::Queues::MeteringWrite,
                 routing_key: 'metering.#'
               },
               {
-                from:        Legion::LLM::Transport::Exchanges::Audit,
+                from:        Legion::Extensions::Llm::Ledger::Transport::Exchanges::Audit,
                 to:          Legion::Extensions::Llm::Ledger::Transport::Queues::AuditPrompts,
                 routing_key: 'audit.prompt.#'
               },
               {
-                from:        Legion::LLM::Transport::Exchanges::Audit,
+                from:        Legion::Extensions::Llm::Ledger::Transport::Exchanges::Audit,
                 to:          Legion::Extensions::Llm::Ledger::Transport::Queues::AuditTools,
                 routing_key: 'audit.tool.#'
               }